+++ lib/src/extended_memory_image_provider.dart
@@ -8,7 +8,15 @@ class ExtendedMemoryImageProvider extends MemoryImage
   const ExtendedMemoryImageProvider(
     Uint8List bytes, {
     double scale = 1.0,
+    this.cacheRawData = false,
   }) : super(bytes, scale: scale);
+
+  /// Whether cache raw data if you need to get raw data directly.
+  /// For example, we need raw image data to edit,
+  /// but [ui.Image.toByteData()] is very slow. So we cache the image
+  /// data here.
+  @override
+  final bool cacheRawData;
 
   @override
   ImageStreamCompleter load(MemoryImage key, DecoderCallback decode) {
@@ -25,4 +33,22 @@ class ExtendedMemoryImageProvider extends MemoryImage
 
   @override
   Uint8List get rawImageData => bytes;
+
+  @override
+  bool operator ==(Object other) {
+    if (other.runtimeType != runtimeType) {
+      return false;
+    }
+    return other is ExtendedMemoryImageProvider &&
+        other.bytes == bytes &&
+        other.scale == scale &&
+        cacheRawData == other.cacheRawData;
+  }
+
+  @override
+  int get hashCode => hashValues(
+        bytes.hashCode,
+        scale,
+        cacheRawData,
+      );
 }