+++ CHANGELOG.md
@@ -1,6 +1,12 @@
-## 2.1.0
+## 3.0.0
 
-* add [ExtendedResizeImage] to support resize image
+* Breaking change:
+  
+  1. we cache raw image pixels as default behavior at previous versions, it's not good for heap memory usage. so add [ExtendedImageProvider.cacheRawData] to support whether should cache the raw image pixels. It's [false] now.
+
+* Improve:
+  
+  1. add [ExtendedResizeImage] to support resize image more convenient.
   
 ## 2.0.2